requirements.txt
eth-brownie==1.16.4
black==21.9b0
isort==5.10.1
